payments refactor
