+++ packages/dashboard/src/features/payments/components/PaymentsSettingsDialog.tsx
@@ -1,18 +1,6 @@
-import { useEffect, useRef, useState, type ReactNode } from 'react';
+import { useState } from 'react';
+import { KeyRound, RefreshCw, Webhook } from 'lucide-react';
 import {
-  CheckCircle2,
-  ExternalLink,
-  Eye,
-  EyeOff,
-  KeyRound,
-  Loader2,
-  RefreshCw,
-  Webhook,
-} from 'lucide-react';
-import {
-  Button,
-  ConfirmDialog,
-  CopyButton,
   MenuDialog,
   MenuDialogBody,
   MenuDialogCloseButton,
@@ -28,128 +16,11 @@ import {
   MenuDialogSideNavTitle,
   MenuDialogTitle,
 } from '@insforge/ui';
-import type {
-  PaymentEnvironment,
-  PaymentProvider,
-  RazorpayKeyConfig,
-  StripeConnection,
-  StripeKeyConfig,
-} from '@insforge/shared-schemas';
-import { usePaymentsConfig } from '#features/payments/hooks/usePaymentsConfig';
-import { usePaymentsSync } from '#features/payments/hooks/usePaymentsSync';
-import { usePaymentsWebhook } from '#features/payments/hooks/usePaymentsWebhook';
-import { useRazorpayConfig } from '#features/payments/hooks/useRazorpayConfig';
-import { useRazorpaySync } from '#features/payments/hooks/useRazorpaySync';
-import {
-  useRazorpayWebhook,
-  useRazorpayWebhookSetup,
-} from '#features/payments/hooks/useRazorpayWebhook';
+import type { PaymentProvider } from '@insforge/shared-schemas';
 import { PaymentProviderSelect, PAYMENT_PROVIDER_LABELS } from './PaymentProviderSelect';
-
-const ENVIRONMENTS: PaymentEnvironment[] = ['test', 'live'];
-type PaymentsSettingsTab = 'keys' | 'webhooks' | 'sync';
-type PaymentKeyConfig = StripeKeyConfig | RazorpayKeyConfig;
-
-const KEY_PREFIX_BY_ENVIRONMENT: Record<PaymentEnvironment, string> = {
-  test: 'sk_test_',
-  live: 'sk_live_',
-};
-
-const RAZORPAY_PREFIX_BY_ENVIRONMENT: Record<PaymentEnvironment, string> = {
-  test: 'rzp_test_',
-  live: 'rzp_live_',
-};
-const RAZORPAY_WEBHOOK_DOCS_URL =
-  'https://razorpay.com/docs/payments/dashboard/account-settings/webhooks/';
-const RAZORPAY_RECOMMENDED_WEBHOOK_EVENTS = [
-  'payment.authorized',
-  'payment.captured',
-  'payment.failed',
-  'order.paid',
-  'refund.created',
-  'refund.processed',
-  'refund.failed',
-  'subscription.created',
-  'subscription.activated',
-  'subscription.charged',
-  'subscription.updated',
-  'subscription.cancelled',
-  'subscription.paused',
-  'subscription.resumed',
-  'subscription.halted',
-  'subscription.completed',
-  'subscription.expired',
-  'invoice.paid',
-  'invoice.expired',
-];
-
-function getConfiguredRazorpayApiKeys(keys: RazorpayKeyConfig[]): RazorpayKeyConfig[] {
-  const environmentsWithSecret = new Set(
-    keys
-      .filter((key) => key.keyType === 'api_secret' && Boolean(key.value))
-      .map((key) => key.environment)
-  );
-
-  return keys.filter(
-    (key) =>
-      key.keyType === 'api_key' && Boolean(key.value) && environmentsWithSecret.has(key.environment)
-  );
-}
-
-function createEmptyEnvironmentValues(): Record<PaymentEnvironment, string> {
-  return { test: '', live: '' };
-}
-
-function getStripeKeyValue(keys: StripeKeyConfig[], environment: PaymentEnvironment): string {
-  return keys.find((key) => key.environment === environment)?.value ?? '';
-}
-
-function getStripeKeyValues(keys: StripeKeyConfig[]): Record<PaymentEnvironment, string> {
-  return {
-    test: getStripeKeyValue(keys, 'test'),
-    live: getStripeKeyValue(keys, 'live'),
-  };
-}
-
-function getRazorpayKeyValue(
-  keys: RazorpayKeyConfig[],
-  environment: PaymentEnvironment,
-  keyType: RazorpayKeyConfig['keyType']
-): string {
-  return (
-    keys.find((key) => key.environment === environment && key.keyType === keyType)?.value ?? ''
-  );
-}
-
-function getRazorpayKeyValues(
-  keys: RazorpayKeyConfig[],
-  keyType: RazorpayKeyConfig['keyType']
-): Record<PaymentEnvironment, string> {
-  return {
-    test: getRazorpayKeyValue(keys, 'test', keyType),
-    live: getRazorpayKeyValue(keys, 'live', keyType),
-  };
-}
-
-function hydrateEnvironmentValues(
-  current: Record<PaymentEnvironment, string>,
-  previousSaved: Record<PaymentEnvironment, string>,
-  nextSaved: Record<PaymentEnvironment, string>
-): Record<PaymentEnvironment, string> {
-  let changed = false;
-  const next = { ...current };
-
-  for (const environment of ENVIRONMENTS) {
-    const canHydrate =
-      current[environment] === '' || current[environment] === previousSaved[environment];
-    if (canHydrate && current[environment] !== nextSaved[environment]) {
-      next[environment] = nextSaved[environment];
-      changed = true;
-    }
-  }
-
-  return changed ? next : current;
-}
+import { StripeSettingsPanel, useStripeSettings } from './StripeSettingsPanel';
+import { RazorpaySettingsPanel, useRazorpaySettings } from './RazorpaySettingsPanel';
+import type { PaymentsSettingsTab } from './PaymentsSettingsShared';
 
 interface PaymentsSettingsDialogProps {
   open: boolean;
@@ -158,1022 +29,24 @@ interface PaymentsSettingsDialogProps {
   setProvider: (provider: PaymentProvider) => void;
 }
 
-interface SettingRowProps {
-  label: string;
-  description?: ReactNode;
-  children: ReactNode;
-  orientation?: 'horizontal' | 'vertical';
-}
-
-function SettingRow({ label, description, children, orientation = 'horizontal' }: SettingRowProps) {
-  if (orientation === 'vertical') {
-    return (
-      <div className="flex w-full flex-col items-start gap-2">
-        <div className="w-full shrink-0">
-          <div className="py-1 flex items-center">
-            <p className="text-sm font-medium leading-5 text-foreground">{label}</p>
-          </div>
-          {description && (
-            <div className="pb-3 text-[13px] leading-[18px] text-muted-foreground">
-              {description}
-            </div>
-          )}
-        </div>
-        <div className="min-w-0 w-full">{children}</div>
-      </div>
-    );
-  }
-
-  return (
-    <div className="flex w-full items-start gap-6">
-      <div className="w-[200px] shrink-0">
-        <div className="py-1.5">
-          <p className="text-sm leading-5 text-foreground">{label}</p>
-        </div>
-        {description && (
-          <div className="pb-2 pt-1 text-[13px] leading-[18px] text-muted-foreground">
-            {description}
-          </div>
-        )}
-      </div>
-      <div className="min-w-0 flex-1">{children}</div>
-    </div>
-  );
-}
-
-function WebhookStatusBadge({ configured }: { configured: boolean }) {
-  if (!configured) {
-    return (
-      <span className="inline-flex items-center rounded-full border border-[var(--alpha-8)] bg-muted px-2 py-0.5 text-xs font-medium text-muted-foreground">
-        Not configured
-      </span>
-    );
-  }
-
-  return (
-    <span className="inline-flex items-center gap-1.5 rounded-full bg-primary/20 px-2 py-0.5 text-xs font-medium text-primary">
-      <CheckCircle2 className="h-3 w-3" />
-      Configured
-    </span>
-  );
-}
-
-function formatWebhookConfiguredAt(value: string | null | undefined) {
-  if (!value) {
-    return 'Not configured';
-  }
-
-  const date = new Date(value);
-  if (Number.isNaN(date.getTime())) {
-    return value;
-  }
-
-  return new Intl.DateTimeFormat(undefined, {
-    dateStyle: 'medium',
-    timeStyle: 'short',
-  }).format(date);
-}
-
-function isPublicHttpsWebhookUrl(value: string) {
-  try {
-    const url = new URL(value);
-    return (
-      url.protocol === 'https:' &&
-      url.hostname !== 'localhost' &&
-      url.hostname !== '127.0.0.1' &&
-      url.hostname !== '::1' &&
-      url.hostname !== '[::1]' &&
-      !url.hostname.startsWith('127.')
-    );
-  } catch {
-    return false;
-  }
-}
-
-interface EnvironmentKeySectionProps {
-  environment: PaymentEnvironment;
-  config?: StripeKeyConfig;
-  savedValue: string;
-  inputValue: string;
-  showKey: boolean;
-  error?: string;
-  isBusy: boolean;
-  onInputChange: (value: string) => void;
-  onToggleShowKey: () => void;
-  onSave: () => void;
-  onRemove: () => void;
-}
-
-function EnvironmentKeySection({
-  environment,
-  config,
-  savedValue,
-  inputValue,
-  showKey,
-  error,
-  isBusy,
-  onInputChange,
-  onToggleShowKey,
-  onSave,
-  onRemove,
-}: EnvironmentKeySectionProps) {
-  const expectedPrefix = KEY_PREFIX_BY_ENVIRONMENT[environment];
-  const environmentLabel = environment === 'test' ? 'Test Mode' : 'Live Mode';
-  const hasSavedValue = Boolean(config?.value);
-  const hasPendingInput = inputValue.trim() !== savedValue.trim();
-
-  return (
-    <SettingRow
-      label={environmentLabel}
-      description={
-        <>
-          Use a Stripe secret key that starts with{' '}
-          <span className="font-mono text-foreground">{expectedPrefix}</span>
-        </>
-      }
-    >
-      <div className="flex flex-col gap-2">
-        <div className="relative min-w-0">
-          <div className="relative">
-            <input
-              type={showKey ? 'text' : 'password'}
-              value={inputValue}
-              onChange={(event) => onInputChange(event.target.value)}
-              placeholder={expectedPrefix}
-              disabled={isBusy}
-              className="h-8 w-full rounded border border-[var(--alpha-12)] bg-[var(--alpha-4)] px-2.5 pr-9 text-sm leading-5 text-foreground placeholder:text-muted-foreground focus:outline-none focus:shadow-[0_0_0_1px_rgb(var(--inverse)),0_0_0_2px_rgb(var(--foreground))] hover:bg-[var(--alpha-4)] disabled:opacity-50"
-            />
-            <button
-              type="button"
-              onClick={onToggleShowKey}
-              className="absolute right-2 top-1/2 -translate-y-1/2 text-muted-foreground hover:text-foreground"
-              aria-label={showKey ? 'Hide key' : 'Show key'}
-              disabled={isBusy}
-            >
-              {showKey ? <EyeOff className="h-4 w-4" /> : <Eye className="h-4 w-4" />}
-            </button>
-          </div>
-        </div>
-
-        {error && <p className="text-xs text-destructive">{error}</p>}
-
-        {(hasSavedValue || hasPendingInput) && (
-          <div className="flex flex-wrap justify-end gap-2">
-            <div className="flex items-center gap-2">
-              {hasSavedValue && (
-                <Button
-                  type="button"
-                  variant="destructive"
-                  size="sm"
-                  onClick={onRemove}
-                  disabled={isBusy}
-                  className="h-7 px-2"
-                >
-                  Remove
-                </Button>
-              )}
-
-              {hasPendingInput && (
-                <Button
-                  type="button"
-                  variant="primary"
-                  size="sm"
-                  onClick={onSave}
-                  disabled={isBusy}
-                  className="h-7 px-2"
-                >
-                  {isBusy ? <Loader2 className="h-3.5 w-3.5 animate-spin" /> : null}
-                  Save
-                </Button>
-              )}
-            </div>
-          </div>
-        )}
-      </div>
-    </SettingRow>
-  );
-}
-
-function DialogSectionDivider() {
-  return (
-    <div className="flex h-5 items-center justify-center">
-      <div className="h-px w-full bg-[var(--alpha-8)]" />
-    </div>
-  );
-}
-
-function StripeKeysTabContent({
-  keys,
-  isLoading,
-  error,
-  isBusy,
-  keyInputs,
-  visibleKeys,
-  errors,
-  onInputChange,
-  onToggleShowKey,
-  onSave,
-  onRemove,
-}: {
-  keys: StripeKeyConfig[];
-  isLoading: boolean;
-  error: unknown;
-  isBusy: boolean;
-  keyInputs: Record<PaymentEnvironment, string>;
-  visibleKeys: Record<PaymentEnvironment, boolean>;
-  errors: Partial<Record<PaymentEnvironment, string>>;
-  onInputChange: (environment: PaymentEnvironment, value: string) => void;
-  onToggleShowKey: (environment: PaymentEnvironment) => void;
-  onSave: (environment: PaymentEnvironment) => void;
-  onRemove: (environment: PaymentEnvironment) => void;
-}) {
-  if (isLoading && !error) {
-    return (
-      <div className="flex min-h-[120px] items-center justify-center gap-2 text-sm text-muted-foreground">
-        <Loader2 className="h-4 w-4 animate-spin" />
-        Loading Stripe key configuration...
-      </div>
-    );
-  }
-
-  if (error) {
-    return (
-      <div className="rounded border border-destructive/20 bg-destructive/10 p-3 text-sm text-destructive">
-        Failed to load Stripe key configuration. Close the dialog and try again.
-      </div>
-    );
-  }
-
-  return (
-    <div className="flex flex-col gap-6">
-      <div>
-        <p className="text-sm leading-6 text-muted-foreground">
-          Configure the Stripe secret keys to use Payments.
-        </p>
-      </div>
-
-      <div className="flex flex-col gap-2">
-        {ENVIRONMENTS.map((environment, index) => (
-          <div key={environment} className="flex flex-col gap-2">
-            <EnvironmentKeySection
-              environment={environment}
-              config={keys.find((key) => key.environment === environment)}
-              savedValue={getStripeKeyValue(keys, environment)}
-              inputValue={keyInputs[environment]}
-              showKey={visibleKeys[environment]}
-              error={errors[environment]}
-              isBusy={isBusy}
-              onInputChange={(value) => onInputChange(environment, value)}
-              onToggleShowKey={() => onToggleShowKey(environment)}
-              onSave={() => onSave(environment)}
-              onRemove={() => onRemove(environment)}
-            />
-            {index < ENVIRONMENTS.length - 1 && <DialogSectionDivider />}
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-}
-
-// RAZORPAY KEYS TAB CONTENT
-function RazorpayKeysTabContent({
-  keys,
-  isLoading,
-  error,
-  isBusy,
-  keyIdInputs,
-  keySecretInputs,
-  visibleKeys,
-  errors,
-  onIdInputChange,
-  onSecretInputChange,
-  onToggleShowKey,
-  onSave,
-  onRemove,
-}: {
-  keys: RazorpayKeyConfig[];
-  isLoading: boolean;
-  error: unknown;
-  isBusy: boolean;
-  keyIdInputs: Record<PaymentEnvironment, string>;
-  keySecretInputs: Record<PaymentEnvironment, string>;
-  visibleKeys: Record<PaymentEnvironment, boolean>;
-  errors: Partial<Record<PaymentEnvironment, string>>;
-  onIdInputChange: (environment: PaymentEnvironment, value: string) => void;
-  onSecretInputChange: (environment: PaymentEnvironment, value: string) => void;
-  onToggleShowKey: (environment: PaymentEnvironment) => void;
-  onSave: (environment: PaymentEnvironment) => void;
-  onRemove: (environment: PaymentEnvironment) => void;
-}) {
-  if (isLoading && !error) {
-    return (
-      <div className="flex min-h-[120px] items-center justify-center gap-2 text-sm text-muted-foreground">
-        <Loader2 className="h-4 w-4 animate-spin" />
-        Loading Razorpay key configuration...
-      </div>
-    );
-  }
-
-  if (error) {
-    return (
-      <div className="rounded border border-destructive/20 bg-destructive/10 p-3 text-sm text-destructive">
-        Failed to load Razorpay key configuration. Close the dialog and try again.
-      </div>
-    );
-  }
-
-  return (
-    <div className="flex flex-col gap-6">
-      <div>
-        <p className="text-sm leading-6 text-muted-foreground">
-          Configure the Razorpay API Keys to use Payments.
-        </p>
-      </div>
-
-      <div className="flex flex-col gap-6">
-        {ENVIRONMENTS.map((environment, index) => {
-          const envIdKey = keys.find(
-            (key) => key.environment === environment && key.keyType === 'api_key'
-          );
-          const envSecretKey = keys.find(
-            (key) => key.environment === environment && key.keyType === 'api_secret'
-          );
-
-          const hasAnyKey = Boolean(envIdKey?.value || envSecretKey?.value);
-          const expectedPrefix = RAZORPAY_PREFIX_BY_ENVIRONMENT[environment];
-          const environmentLabel = environment === 'test' ? 'Test Mode' : 'Live Mode';
-          const savedKeyId = envIdKey?.value ?? '';
-          const savedKeySecret = envSecretKey?.value ?? '';
-          const hasPendingInput =
-            keyIdInputs[environment].trim() !== savedKeyId.trim() ||
-            keySecretInputs[environment].trim() !== savedKeySecret.trim();
-          const keyIdInputId = `razorpay-${environment}-key-id`;
-          const keySecretInputId = `razorpay-${environment}-key-secret`;
-
-          return (
-            <div key={environment} className="flex flex-col gap-2">
-              <SettingRow
-                label={environmentLabel}
-                description={
-                  <>
-                    Use a Razorpay Key ID that starts with{' '}
-                    <span className="font-mono text-foreground">{expectedPrefix}</span> and its
-                    matching Key Secret.
-                  </>
-                }
-              >
-                <div className="flex flex-col gap-2">
-                  <div className="flex flex-col gap-2">
-                    <div className="grid grid-cols-[96px_minmax(0,1fr)] items-center gap-3">
-                      <label
-                        htmlFor={keyIdInputId}
-                        className="text-xs font-medium text-muted-foreground"
-                      >
-                        Key ID
-                      </label>
-                      <input
-                        id={keyIdInputId}
-                        type="text"
-                        value={keyIdInputs[environment]}
-                        onChange={(event) => onIdInputChange(environment, event.target.value)}
-                        placeholder={`${expectedPrefix}...`}
-                        disabled={isBusy}
-                        className="h-8 w-full rounded border border-[var(--alpha-12)] bg-[var(--alpha-4)] px-2.5 text-sm leading-5 text-foreground placeholder:text-muted-foreground focus:outline-none focus:shadow-[0_0_0_1px_rgb(var(--inverse)),0_0_0_2px_rgb(var(--foreground))] hover:bg-[var(--alpha-4)] disabled:opacity-50"
-                      />
-                    </div>
-                    <div className="grid grid-cols-[96px_minmax(0,1fr)] items-center gap-3">
-                      <label
-                        htmlFor={keySecretInputId}
-                        className="text-xs font-medium text-muted-foreground"
-                      >
-                        Key Secret
-                      </label>
-                      <div className="relative">
-                        <input
-                          id={keySecretInputId}
-                          type={visibleKeys[environment] ? 'text' : 'password'}
-                          value={keySecretInputs[environment]}
-                          onChange={(event) => onSecretInputChange(environment, event.target.value)}
-                          placeholder="Enter key secret"
-                          disabled={isBusy}
-                          className="h-8 w-full rounded border border-[var(--alpha-12)] bg-[var(--alpha-4)] px-2.5 pr-9 text-sm leading-5 text-foreground placeholder:text-muted-foreground focus:outline-none focus:shadow-[0_0_0_1px_rgb(var(--inverse)),0_0_0_2px_rgb(var(--foreground))] hover:bg-[var(--alpha-4)] disabled:opacity-50"
-                        />
-                        <button
-                          type="button"
-                          onClick={() => onToggleShowKey(environment)}
-                          className="absolute right-2 top-1/2 -translate-y-1/2 text-muted-foreground hover:text-foreground"
-                          aria-label={
-                            visibleKeys[environment] ? 'Hide key secret' : 'Show key secret'
-                          }
-                          disabled={isBusy}
-                        >
-                          {visibleKeys[environment] ? (
-                            <EyeOff className="h-4 w-4" />
-                          ) : (
-                            <Eye className="h-4 w-4" />
-                          )}
-                        </button>
-                      </div>
-                    </div>
-                  </div>
-
-                  {errors[environment] && (
-                    <p className="text-xs text-destructive">{errors[environment]}</p>
-                  )}
-
-                  {(hasAnyKey || hasPendingInput) && (
-                    <div className="mt-2 flex flex-wrap justify-end gap-2">
-                      <div className="flex items-center gap-2">
-                        {hasAnyKey && (
-                          <Button
-                            type="button"
-                            variant="destructive"
-                            size="sm"
-                            onClick={() => onRemove(environment)}
-                            disabled={isBusy}
-                            className="h-7 px-2"
-                          >
-                            Remove
-                          </Button>
-                        )}
-
-                        {hasPendingInput && (
-                          <Button
-                            type="button"
-                            variant="primary"
-                            size="sm"
-                            onClick={() => onSave(environment)}
-                            disabled={isBusy}
-                            className="h-7 px-2"
-                          >
-                            {isBusy ? <Loader2 className="h-3.5 w-3.5 animate-spin" /> : null}
-                            Save
-                          </Button>
-                        )}
-                      </div>
-                    </div>
-                  )}
-                </div>
-              </SettingRow>
-              {index < ENVIRONMENTS.length - 1 && <DialogSectionDivider />}
-            </div>
-          );
-        })}
-      </div>
-    </div>
-  );
-}
-
-function SyncTabContent({
-  isLoading,
-  error,
-  configuredKeys,
-  syncPayments,
-  onSync,
-  provider = 'stripe',
-}: {
-  isLoading: boolean;
-  error: unknown;
-  configuredKeys: PaymentKeyConfig[];
-  syncPayments:
-    | ReturnType<typeof usePaymentsSync>['syncPayments']
-    | ReturnType<typeof useRazorpaySync>['syncPayments'];
-  onSync: () => void;
-  provider?: 'stripe' | 'razorpay';
-}) {
-  const providerName = provider === 'stripe' ? 'Stripe' : 'Razorpay';
-  const syncDescription =
-    provider === 'stripe'
-      ? 'Force a manual sync of Stripe products, prices, customers, and active subscriptions. Normally, Stripe events sync automatically via webhooks.'
-      : 'Force a manual sync of Razorpay items, plans, customers, subscriptions, invoices, and payments. Normally, Razorpay events sync automatically via webhooks.';
-
-  if (isLoading && !error) {
-    return (
-      <div className="flex min-h-[120px] items-center justify-center gap-2 text-sm text-muted-foreground">
-        <Loader2 className="h-4 w-4 animate-spin" />
-        Loading {providerName} key configuration...
-      </div>
-    );
-  }
-
-  if (error) {
-    return (
-      <div className="rounded border border-destructive/20 bg-destructive/10 p-3 text-sm text-destructive">
-        Failed to load {providerName} key configuration. Close the dialog and try again.
-      </div>
-    );
-  }
-
-  return (
-    <div className="flex flex-col gap-6">
-      <div>
-        <p className="text-sm leading-6 text-muted-foreground">{syncDescription}</p>
-      </div>
-
-      <div className="rounded border border-[var(--alpha-8)] bg-muted/40 p-4">
-        <div className="flex items-start justify-between gap-4">
-          <div className="min-w-0">
-            <p className="text-sm font-medium text-foreground">Sync all configured environments</p>
-            <p className="mt-1 text-[13px] leading-5 text-muted-foreground">
-              {configuredKeys.length > 0
-                ? `Configured: ${configuredKeys
-                    .map((key) => (key.environment === 'test' ? 'Test' : 'Live'))
-                    .join(', ')}`
-                : `Configure a ${providerName} test or live key before syncing.`}
-            </p>
-          </div>
-          <Button
-            type="button"
-            size="lg"
-            onClick={onSync}
-            disabled={syncPayments.isPending || configuredKeys.length === 0}
-            className="h-9 shrink-0"
-          >
-            {syncPayments.isPending ? (
-              <Loader2 className="h-4 w-4 animate-spin" />
-            ) : (
-              <RefreshCw className="h-4 w-4" />
-            )}
-            Sync Payments
-          </Button>
-        </div>
-      </div>
-
-      {syncPayments.error && (
-        <div className="rounded border border-destructive/20 bg-destructive/10 p-3 text-sm text-destructive">
-          {syncPayments.error instanceof Error
-            ? syncPayments.error.message
-            : `Failed to sync ${providerName} payments.`}
-        </div>
-      )}
-    </div>
-  );
-}
-
-function StripeWebhooksTabContent({
-  keys,
-  connections,
-  isLoading,
-  isLoadingWebhooks,
-  error,
-  webhooksError,
-  configureWebhook,
-  isBusy,
-  onConfigure,
-}: {
-  keys: StripeKeyConfig[];
-  connections: StripeConnection[];
-  isLoading: boolean;
-  isLoadingWebhooks: boolean;
-  error: unknown;
-  webhooksError: unknown;
-  configureWebhook: ReturnType<typeof usePaymentsWebhook>['configureWebhook'];
-  isBusy: boolean;
-  onConfigure: (environment: PaymentEnvironment) => void;
-}) {
-  if ((isLoading || isLoadingWebhooks) && !error && !webhooksError) {
-    return (
-      <div className="flex min-h-[120px] items-center justify-center gap-2 text-sm text-muted-foreground">
-        <Loader2 className="h-4 w-4 animate-spin" />
-        Loading Stripe webhook configuration...
-      </div>
-    );
-  }
-
-  if (error || webhooksError) {
-    return (
-      <div className="rounded border border-destructive/20 bg-destructive/10 p-3 text-sm text-destructive">
-        Failed to load Stripe webhook configuration. Close the dialog and try again.
-      </div>
-    );
-  }
-
-  return (
-    <div className="flex flex-col gap-6">
-      <div>
-        <p className="text-sm leading-6 text-muted-foreground">
-          Configure Stripe webhook endpoints for customer, transaction, and subscription updates.
-        </p>
-      </div>
-
-      {ENVIRONMENTS.map((environment) => (
-        <StripeWebhookEnvironmentSection
-          key={environment}
-          environment={environment}
-          config={keys.find((key) => key.environment === environment)}
-          connection={connections.find((connection) => connection.environment === environment)}
-          isConfiguring={configureWebhook.isPending && configureWebhook.variables === environment}
-          isBusy={isBusy}
-          onConfigure={() => onConfigure(environment)}
-        />
-      ))}
-
-      {configureWebhook.error && (
-        <div className="rounded border border-destructive/20 bg-destructive/10 p-3 text-sm text-destructive">
-          {configureWebhook.error instanceof Error
-            ? configureWebhook.error.message
-            : 'Failed to configure Stripe webhook.'}
-        </div>
-      )}
-    </div>
-  );
-}
-
-interface StripeWebhookEnvironmentSectionProps {
-  environment: PaymentEnvironment;
-  config?: StripeKeyConfig;
-  connection?: StripeConnection;
-  isConfiguring: boolean;
-  isBusy: boolean;
-  onConfigure: () => void;
-}
-
-function StripeWebhookEnvironmentSection({
-  environment,
-  config,
-  connection,
-  isConfiguring,
-  isBusy,
-  onConfigure,
-}: StripeWebhookEnvironmentSectionProps) {
-  const environmentLabel = environment === 'test' ? 'Test mode' : 'Live mode';
-  const keyName = environment === 'test' ? 'STRIPE_TEST_SECRET_KEY' : 'STRIPE_LIVE_SECRET_KEY';
-
-  const isKeyConfigured = !!config?.value;
-  const webhookEndpointUrl = connection?.webhookEndpointUrl ?? null;
-  const webhookEndpointId = connection?.webhookEndpointId ?? null;
-  const isWebhookConfigured = !!webhookEndpointId && !!webhookEndpointUrl;
-
-  return (
-    <SettingRow
-      orientation="vertical"
-      label={environmentLabel}
-      description={
-        isKeyConfigured
-          ? 'InsForge creates and stores a Stripe webhook signing secret for this environment.'
-          : `Configure ${keyName} before creating the webhook.`
-      }
-    >
-      <div className="flex flex-col gap-3">
-        <div className="flex flex-wrap items-center gap-2">
-          <WebhookStatusBadge configured={isWebhookConfigured} />
-          {connection?.webhookConfiguredAt && (
-            <span className="text-xs text-muted-foreground">
-              {formatWebhookConfiguredAt(connection.webhookConfiguredAt)}
-            </span>
-          )}
-        </div>
-
-        <div className="rounded border border-[var(--alpha-8)] bg-muted/40 p-3">
-          {isWebhookConfigured ? (
-            <div className="grid gap-2 text-xs">
-              <div className="grid grid-cols-[96px_minmax(0,1fr)] gap-3">
-                <span className="text-muted-foreground">Endpoint</span>
-                <span className="min-w-0 truncate font-mono text-foreground">
-                  {webhookEndpointUrl}
-                </span>
-              </div>
-              <div className="grid grid-cols-[96px_minmax(0,1fr)] gap-3">
-                <span className="text-muted-foreground">Stripe ID</span>
-                <span className="min-w-0 truncate font-mono text-foreground">
-                  {webhookEndpointId}
-                </span>
-              </div>
-              <div className="grid grid-cols-[96px_minmax(0,1fr)] gap-3">
-                <span className="text-muted-foreground">Secret</span>
-                <span className="text-foreground">Stored in InsForge secret store</span>
-              </div>
-            </div>
-          ) : (
-            <p className="text-xs leading-5 text-muted-foreground">
-              {isKeyConfigured
-                ? 'No managed Stripe webhook is configured yet. Create one when your backend has a public API URL.'
-                : 'Webhook setup uses the saved Stripe API key, so the key must be configured first.'}
-            </p>
-          )}
-        </div>
-
-        <div className="flex justify-end">
-          <Button
-            type="button"
-            size="lg"
-            onClick={onConfigure}
-            disabled={!isKeyConfigured || isBusy}
-            className="h-9 shrink-0"
-          >
-            {isConfiguring ? (
-              <Loader2 className="h-4 w-4 animate-spin" />
-            ) : (
-              <Webhook className="h-4 w-4" />
-            )}
-            {isWebhookConfigured ? 'Reconfigure webhook' : 'Configure webhook'}
-          </Button>
-        </div>
-      </div>
-    </SettingRow>
-  );
-}
-
-function RazorpayWebhooksTabContent({
-  keys,
-  rotateWebhookSecret,
-  isBusy,
-  onGoToKeys,
-}: {
-  keys: RazorpayKeyConfig[];
-  rotateWebhookSecret: ReturnType<typeof useRazorpayWebhook>['rotateWebhookSecret'];
-  isBusy: boolean;
-  onGoToKeys: () => void;
-}) {
-  return (
-    <div className="flex flex-col gap-6">
-      <div>
-        <p className="text-sm leading-6 text-muted-foreground">
-          Razorpay webhooks must be added manually in the Razorpay Dashboard.
-        </p>
-      </div>
-
-      {ENVIRONMENTS.map((environment, index) => (
-        <div key={environment} className="flex flex-col gap-2">
-          <RazorpayWebhookEnvironmentSection
-            environment={environment}
-            keys={keys}
-            rotateWebhookSecret={rotateWebhookSecret}
-            isBusy={isBusy}
-            onGoToKeys={onGoToKeys}
-          />
-          {index < ENVIRONMENTS.length - 1 && <DialogSectionDivider />}
-        </div>
-      ))}
-
-      <RazorpayWebhookManualSetupGuidance />
-    </div>
-  );
-}
-
-function RazorpayWebhookManualSetupGuidance() {
-  return (
-    <div className="flex flex-col gap-4">
-      <div className="rounded border border-[var(--alpha-8)] p-3">
-        <div className="flex items-center justify-between gap-3">
-          <p className="text-sm font-medium text-foreground">Manual setup steps</p>
-          <a
-            href={RAZORPAY_WEBHOOK_DOCS_URL}
-            target="_blank"
-            rel="noreferrer"
-            className="inline-flex items-center gap-1 text-xs font-medium text-primary hover:underline"
-          >
-            Razorpay docs
-            <ExternalLink className="h-3.5 w-3.5" />
-          </a>
-        </div>
-        <ol className="mt-3 list-decimal space-y-1 pl-4 text-xs leading-5 text-muted-foreground">
-          <li>Open Razorpay Dashboard and go to Accounts &amp; Settings → Webhooks.</li>
-          <li>Add a webhook for each environment you enable.</li>
-          <li>Paste the matching environment&apos;s Webhook URL and Webhook Secret above.</li>
-          <li>Select the Active Events listed below.</li>
-          <li>Save the webhook, then make a test payment to verify delivery.</li>
-        </ol>
-      </div>
-
-      <div className="rounded border border-[var(--alpha-8)] p-3">
-        <p className="text-sm font-medium text-foreground">Active Events to select</p>
-        <div className="mt-2 flex flex-wrap gap-1.5">
-          {RAZORPAY_RECOMMENDED_WEBHOOK_EVENTS.map((event) => (
-            <span
-              key={event}
-              className="rounded border border-[var(--alpha-8)] bg-muted px-1.5 py-0.5 font-mono text-[11px] text-muted-foreground"
-            >
-              {event}
-            </span>
-          ))}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function RazorpayWebhookEnvironmentSection({
-  environment,
-  keys,
-  rotateWebhookSecret,
-  isBusy,
-  onGoToKeys,
-}: {
-  environment: PaymentEnvironment;
-  keys: RazorpayKeyConfig[];
-  rotateWebhookSecret: ReturnType<typeof useRazorpayWebhook>['rotateWebhookSecret'];
-  isBusy: boolean;
-  onGoToKeys: () => void;
-}) {
-  const [isRotateConfirmOpen, setIsRotateConfirmOpen] = useState(false);
-  const environmentLabel = environment === 'test' ? 'Test mode' : 'Live mode';
-  const hasKeyId = keys.some(
-    (key) => key.environment === environment && key.keyType === 'api_key' && Boolean(key.value)
-  );
-  const hasKeySecret = keys.some(
-    (key) => key.environment === environment && key.keyType === 'api_secret' && Boolean(key.value)
-  );
-  const isKeyConfigured = hasKeyId && hasKeySecret;
-  const setupQuery = useRazorpayWebhookSetup(environment, isKeyConfigured);
-  const setup = setupQuery.data ?? null;
-  const isWebhookUrlPublic = setup ? isPublicHttpsWebhookUrl(setup.webhookUrl) : true;
-  const isRotating = rotateWebhookSecret.isPending && rotateWebhookSecret.variables === environment;
-
-  return (
-    <SettingRow
-      orientation="vertical"
-      label={environmentLabel}
-      description={
-        isKeyConfigured
-          ? 'Copy these values into the Razorpay Dashboard for this environment.'
-          : 'Configure Razorpay keys first.'
-      }
-    >
-      {!isKeyConfigured ? (
-        <div className="rounded border border-[var(--alpha-8)] bg-muted/40 p-4">
-          <p className="text-sm text-muted-foreground">Configure Razorpay keys first.</p>
-          <Button type="button" size="sm" className="mt-3 h-8" onClick={onGoToKeys}>
-            Connection Keys
-          </Button>
-        </div>
-      ) : setupQuery.isLoading ? (
-        <div className="flex min-h-[96px] items-center justify-center gap-2 text-sm text-muted-foreground">
-          <Loader2 className="h-4 w-4 animate-spin" />
-          Loading Razorpay webhook setup values...
-        </div>
-      ) : setupQuery.error || !setup ? (
-        <div className="rounded border border-destructive/20 bg-destructive/10 p-3 text-sm text-destructive">
-          Failed to load Razorpay webhook setup values. Close the dialog and try again.
-        </div>
-      ) : (
-        <div className="flex flex-col gap-4">
-          <div className="rounded border border-[var(--alpha-8)] bg-muted/40 p-3">
-            <div className="grid gap-3 text-xs">
-              <div className="grid grid-cols-[112px_minmax(0,1fr)_auto] items-center gap-3">
-                <span className="text-muted-foreground">Webhook URL</span>
-                <span className="min-w-0 break-all font-mono text-foreground">
-                  {setup.webhookUrl}
-                </span>
-                <CopyButton text={setup.webhookUrl} showText={false} />
-              </div>
-              {!isWebhookUrlPublic && (
-                <div className="rounded border border-destructive/20 bg-destructive/10 p-2 text-xs text-destructive">
-                  Razorpay can only deliver webhooks to a public HTTPS URL.
-                </div>
-              )}
-              <div className="grid grid-cols-[112px_minmax(0,1fr)_auto_auto] items-center gap-3">
-                <span className="text-muted-foreground">Webhook Secret</span>
-                <span className="min-w-0 break-all font-mono text-foreground">
-                  {setup.webhookSecret}
-                </span>
-                <CopyButton text={setup.webhookSecret} showText={false} />
-                <Button
-                  type="button"
-                  variant="secondary"
-                  size="sm"
-                  className="h-7 px-2"
-                  disabled={isBusy}
-                  onClick={() => setIsRotateConfirmOpen(true)}
-                >
-                  {isRotating ? <Loader2 className="h-3.5 w-3.5 animate-spin" /> : null}
-                  Rotate
-                </Button>
-              </div>
-            </div>
-          </div>
-
-          <ConfirmDialog
-            open={isRotateConfirmOpen}
-            onOpenChange={setIsRotateConfirmOpen}
-            title="Rotate Razorpay webhook secret?"
-            description="Rotating the webhook secret will break existing Razorpay webhook deliveries until you update the secret in Razorpay Dashboard."
-            cancelText="Cancel"
-            confirmText="Rotate"
-            destructive
-            isLoading={isRotating}
-            onConfirm={async () => {
-              await rotateWebhookSecret.mutateAsync(environment);
-            }}
-          />
-        </div>
-      )}
-    </SettingRow>
-  );
-}
-
 export function PaymentsSettingsDialog({
   open,
   onOpenChange,
   provider,
   setProvider,
 }: PaymentsSettingsDialogProps) {
-  const { keys, isLoading, error, saveKey, removeKey } = usePaymentsConfig();
-  const { syncPayments } = usePaymentsSync();
-  const {
-    connections,
-    isLoading: isLoadingWebhooks,
-    error: webhooksError,
-    configureWebhook,
-  } = usePaymentsWebhook();
-
-  // Razorpay hooks
-  const {
-    keys: rzpKeys,
-    isLoading: rzpIsLoading,
-    error: rzpError,
-    saveKey: rzpSaveKey,
-    removeKey: rzpRemoveKey,
-  } = useRazorpayConfig();
-  const { syncPayments: rzpSyncPayments } = useRazorpaySync();
-  const { rotateWebhookSecret } = useRazorpayWebhook();
+  const stripe = useStripeSettings(open);
+  const razorpay = useRazorpaySettings(open);
 
   const [activeTab, setActiveTab] = useState<PaymentsSettingsTab>('keys');
 
-  // Stripe state
-  const [keyInputs, setKeyInputs] = useState<Record<PaymentEnvironment, string>>(
-    createEmptyEnvironmentValues
-  );
-  const [visibleKeys, setVisibleKeys] = useState<Record<PaymentEnvironment, boolean>>({
-    test: false,
-    live: false,
-  });
-  const [errors, setErrors] = useState<Partial<Record<PaymentEnvironment, string>>>({});
-  const previousSavedStripeKeys = useRef<Record<PaymentEnvironment, string>>(
-    createEmptyEnvironmentValues()
-  );
-
-  // Razorpay state
-  const [rzpKeyIdInputs, setRzpKeyIdInputs] = useState<Record<PaymentEnvironment, string>>(
-    createEmptyEnvironmentValues
-  );
-  const [rzpKeySecretInputs, setRzpKeySecretInputs] = useState<Record<PaymentEnvironment, string>>(
-    createEmptyEnvironmentValues
-  );
-  const [rzpVisibleKeys, setRzpVisibleKeys] = useState<Record<PaymentEnvironment, boolean>>({
-    test: false,
-    live: false,
-  });
-  const [rzpErrors, setRzpErrors] = useState<Partial<Record<PaymentEnvironment, string>>>({});
-  const previousSavedRazorpayKeyIds = useRef<Record<PaymentEnvironment, string>>(
-    createEmptyEnvironmentValues()
-  );
-  const previousSavedRazorpayKeySecrets = useRef<Record<PaymentEnvironment, string>>(
-    createEmptyEnvironmentValues()
-  );
-
-  const isBusy =
-    saveKey.isPending ||
-    removeKey.isPending ||
-    syncPayments.isPending ||
-    configureWebhook.isPending ||
-    rzpSaveKey.isPending ||
-    rzpRemoveKey.isPending ||
-    rzpSyncPayments.isPending ||
-    rotateWebhookSecret.isPending;
-
+  // A pending mutation on either provider locks the whole dialog so a
+  // provider switch mid-save can't fire conflicting actions.
+  const isBusy = stripe.isPending || razorpay.isPending;
   const canClose = !isBusy;
-  const configuredKeys = keys.filter((key) => Boolean(key.value));
-  const configuredRzpKeys = getConfiguredRazorpayApiKeys(rzpKeys);
   const title =
     activeTab === 'keys' ? 'Connection Keys' : activeTab === 'webhooks' ? 'Webhooks' : 'Sync';
   const providerName = PAYMENT_PROVIDER_LABELS[provider];
-
-  useEffect(() => {
-    if (!open) {
-      return;
-    }
-
-    const nextSaved = getStripeKeyValues(keys);
-    setKeyInputs((current) =>
-      hydrateEnvironmentValues(current, previousSavedStripeKeys.current, nextSaved)
-    );
-    previousSavedStripeKeys.current = nextSaved;
-  }, [keys, open]);
-
-  useEffect(() => {
-    if (!open) {
-      return;
-    }
-
-    const nextSavedKeyIds = getRazorpayKeyValues(rzpKeys, 'api_key');
-    const nextSavedKeySecrets = getRazorpayKeyValues(rzpKeys, 'api_secret');
-
-    setRzpKeyIdInputs((current) =>
-      hydrateEnvironmentValues(current, previousSavedRazorpayKeyIds.current, nextSavedKeyIds)
-    );
-    setRzpKeySecretInputs((current) =>
-      hydrateEnvironmentValues(
-        current,
-        previousSavedRazorpayKeySecrets.current,
-        nextSavedKeySecrets
-      )
-    );
-
-    previousSavedRazorpayKeyIds.current = nextSavedKeyIds;
-    previousSavedRazorpayKeySecrets.current = nextSavedKeySecrets;
-  }, [open, rzpKeys]);
 
   const handleOpenChange = (nextOpen: boolean) => {
     if (!canClose) {
@@ -1181,119 +54,12 @@ export function PaymentsSettingsDialog({
     }
 
     if (!nextOpen) {
-      setKeyInputs(createEmptyEnvironmentValues());
-      setVisibleKeys({ test: false, live: false });
-      setErrors({});
-      setRzpKeyIdInputs(createEmptyEnvironmentValues());
-      setRzpKeySecretInputs(createEmptyEnvironmentValues());
-      setRzpVisibleKeys({ test: false, live: false });
-      setRzpErrors({});
-
-      saveKey.reset();
-      removeKey.reset();
-      syncPayments.reset();
-      configureWebhook.reset();
-
-      rzpSaveKey.reset();
-      rzpRemoveKey.reset();
-      rzpSyncPayments.reset();
-      rotateWebhookSecret.reset();
-
+      stripe.reset();
+      razorpay.reset();
       setActiveTab('keys');
     }
 
     onOpenChange(nextOpen);
-  };
-
-  const handleSave = async (environment: PaymentEnvironment) => {
-    const secretKey = keyInputs[environment].trim();
-    const expectedPrefix = KEY_PREFIX_BY_ENVIRONMENT[environment];
-
-    if (!secretKey) {
-      setErrors((current) => ({ ...current, [environment]: 'Please enter a Stripe secret key.' }));
-      return;
-    }
-
-    if (!secretKey.startsWith(expectedPrefix)) {
-      setErrors((current) => ({
-        ...current,
-        [environment]: `The ${environment} key must start with ${expectedPrefix}.`,
-      }));
-      return;
-    }
-
-    setErrors((current) => ({ ...current, [environment]: undefined }));
-
-    try {
-      await saveKey.mutateAsync({ environment, secretKey });
-    } catch (err) {
-      setErrors((current) => ({
-        ...current,
-        [environment]: err instanceof Error ? err.message : 'Failed to save Stripe key.',
-      }));
-    }
-  };
-
-  const handleRzpSave = (environment: PaymentEnvironment) => {
-    const keyId = rzpKeyIdInputs[environment].trim();
-    const secretKey = rzpKeySecretInputs[environment].trim();
-    const expectedPrefix = RAZORPAY_PREFIX_BY_ENVIRONMENT[environment];
-
-    if (!keyId || !secretKey) {
-      setRzpErrors((current) => ({
-        ...current,
-        [environment]: 'Please enter both Key ID and Key Secret.',
-      }));
-      return;
-    }
-
-    if (!keyId.startsWith(expectedPrefix)) {
-      setRzpErrors((current) => ({
-        ...current,
-        [environment]: `Razorpay Key ID must start with ${expectedPrefix}`,
-      }));
-      return;
-    }
-
-    setRzpErrors((current) => ({ ...current, [environment]: undefined }));
-    rzpSaveKey.mutate({ environment, keyId, keySecret: secretKey });
-  };
-
-  const handleRemove = async (environment: PaymentEnvironment) => {
-    setErrors((current) => ({ ...current, [environment]: undefined }));
-    try {
-      await removeKey.mutateAsync(environment);
-      setKeyInputs((current) => ({ ...current, [environment]: '' }));
-      setVisibleKeys((current) => ({ ...current, [environment]: false }));
-    } catch (err) {
-      setErrors((current) => ({
-        ...current,
-        [environment]: err instanceof Error ? err.message : 'Failed to remove Stripe key.',
-      }));
-    }
-  };
-
-  const handleRzpRemove = async (environment: PaymentEnvironment) => {
-    setRzpErrors((current) => ({ ...current, [environment]: undefined }));
-    try {
-      await rzpRemoveKey.mutateAsync(environment);
-      setRzpKeyIdInputs((current) => ({ ...current, [environment]: '' }));
-      setRzpKeySecretInputs((current) => ({ ...current, [environment]: '' }));
-      setRzpVisibleKeys((current) => ({ ...current, [environment]: false }));
-    } catch (err) {
-      setRzpErrors((current) => ({
-        ...current,
-        [environment]: err instanceof Error ? err.message : 'Failed to remove Razorpay keys.',
-      }));
-    }
-  };
-
-  const handleConfigureWebhook = async (environment: PaymentEnvironment) => {
-    try {
-      await configureWebhook.mutateAsync(environment);
-    } catch {
-      // The mutation owns toast/error state.
-    }
   };
 
   return (
@@ -1347,92 +113,14 @@ export function PaymentsSettingsDialog({
           </MenuDialogHeader>
 
           <MenuDialogBody>
-            {activeTab === 'keys' ? (
-              provider === 'stripe' ? (
-                <StripeKeysTabContent
-                  keys={keys}
-                  isLoading={isLoading}
-                  error={error}
-                  isBusy={isBusy}
-                  keyInputs={keyInputs}
-                  visibleKeys={visibleKeys}
-                  errors={errors}
-                  onInputChange={(environment, value) =>
-                    setKeyInputs((current) => ({ ...current, [environment]: value }))
-                  }
-                  onToggleShowKey={(environment) =>
-                    setVisibleKeys((current) => ({
-                      ...current,
-                      [environment]: !current[environment],
-                    }))
-                  }
-                  onSave={(environment) => void handleSave(environment)}
-                  onRemove={(environment) => void handleRemove(environment)}
-                />
-              ) : (
-                <RazorpayKeysTabContent
-                  keys={rzpKeys}
-                  isLoading={rzpIsLoading}
-                  error={rzpError}
-                  isBusy={isBusy}
-                  keyIdInputs={rzpKeyIdInputs}
-                  keySecretInputs={rzpKeySecretInputs}
-                  visibleKeys={rzpVisibleKeys}
-                  errors={rzpErrors}
-                  onIdInputChange={(environment, value) =>
-                    setRzpKeyIdInputs((current) => ({ ...current, [environment]: value }))
-                  }
-                  onSecretInputChange={(environment, value) =>
-                    setRzpKeySecretInputs((current) => ({ ...current, [environment]: value }))
-                  }
-                  onToggleShowKey={(environment) =>
-                    setRzpVisibleKeys((current) => ({
-                      ...current,
-                      [environment]: !current[environment],
-                    }))
-                  }
-                  onSave={(environment) => void handleRzpSave(environment)}
-                  onRemove={(environment) => void handleRzpRemove(environment)}
-                />
-              )
-            ) : activeTab === 'webhooks' ? (
-              provider === 'stripe' ? (
-                <StripeWebhooksTabContent
-                  keys={keys}
-                  connections={connections}
-                  isLoading={isLoading}
-                  isLoadingWebhooks={isLoadingWebhooks}
-                  error={error}
-                  webhooksError={webhooksError}
-                  configureWebhook={configureWebhook}
-                  isBusy={isBusy}
-                  onConfigure={(environment) => void handleConfigureWebhook(environment)}
-                />
-              ) : (
-                <RazorpayWebhooksTabContent
-                  keys={rzpKeys}
-                  rotateWebhookSecret={rotateWebhookSecret}
-                  isBusy={isBusy}
-                  onGoToKeys={() => setActiveTab('keys')}
-                />
-              )
-            ) : provider === 'stripe' ? (
-              <SyncTabContent
-                isLoading={isLoading}
-                error={error}
-                configuredKeys={configuredKeys}
-                syncPayments={syncPayments}
-                onSync={() => void syncPayments.mutateAsync({ environment: 'all' })}
-                provider="stripe"
-              />
+            {provider === 'stripe' ? (
+              <StripeSettingsPanel activeTab={activeTab} state={stripe} isBusy={isBusy} />
             ) : (
-              <SyncTabContent
-                isLoading={rzpIsLoading}
-                error={rzpError}
-                configuredKeys={configuredRzpKeys}
-                syncPayments={rzpSyncPayments}
-                onSync={() => void rzpSyncPayments.mutateAsync({ environment: 'all' })}
-                provider="razorpay"
+              <RazorpaySettingsPanel
+                activeTab={activeTab}
+                state={razorpay}
+                isBusy={isBusy}
+                onGoToKeys={() => setActiveTab('keys')}
               />
             )}
           </MenuDialogBody>
