--- packages/dashboard/src/features/payments/hooks/usePaymentsConfig.ts
@@ -1,65 +1,0 @@
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
-import type {
-  GetStripeConfigResponse,
-  StripeEnvironment,
-  UpsertStripeConfigRequest,
-} from '@insforge/shared-schemas';
-import { paymentsService } from '#features/payments/services/payments.service';
-import { useToast } from '#lib/hooks/useToast';
-
-const PAYMENTS_CONFIG_QUERY_KEY = ['payments', 'config'];
-
-export function usePaymentsConfig() {
-  const queryClient = useQueryClient();
-  const { showToast } = useToast();
-
-  const { data, isLoading, error } = useQuery<GetStripeConfigResponse>({
-    queryKey: PAYMENTS_CONFIG_QUERY_KEY,
-    queryFn: () => paymentsService.getConfig(),
-    staleTime: 30 * 1000,
-  });
-
-  const saveKey = useMutation({
-    mutationFn: (input: UpsertStripeConfigRequest) => paymentsService.upsertConfig(input),
-    onSuccess: async () => {
-      await Promise.all([
-        queryClient.invalidateQueries({ queryKey: PAYMENTS_CONFIG_QUERY_KEY }),
-        queryClient.invalidateQueries({ queryKey: ['payments', 'status'] }),
-        queryClient.invalidateQueries({ queryKey: ['payments', 'stripe', 'catalog'] }),
-        queryClient.invalidateQueries({ queryKey: ['payments', 'stripe', 'customers'] }),
-        queryClient.invalidateQueries({ queryKey: ['payments', 'stripe', 'subscriptions'] }),
-        queryClient.invalidateQueries({ queryKey: ['payments', 'stripe', 'transactions'] }),
-      ]);
-      showToast('Stripe key saved successfully', 'success');
-    },
-    onError: (err: Error) => {
-      showToast(err.message || 'Failed to save Stripe key', 'error');
-    },
-  });
-
-  const removeKey = useMutation({
-    mutationFn: (environment: StripeEnvironment) => paymentsService.removeConfig(environment),
-    onSuccess: async () => {
-      await Promise.all([
-        queryClient.invalidateQueries({ queryKey: PAYMENTS_CONFIG_QUERY_KEY }),
-        queryClient.invalidateQueries({ queryKey: ['payments', 'status'] }),
-        queryClient.invalidateQueries({ queryKey: ['payments', 'stripe', 'catalog'] }),
-        queryClient.invalidateQueries({ queryKey: ['payments', 'stripe', 'customers'] }),
-        queryClient.invalidateQueries({ queryKey: ['payments', 'stripe', 'subscriptions'] }),
-        queryClient.invalidateQueries({ queryKey: ['payments', 'stripe', 'transactions'] }),
-      ]);
-      showToast('Stripe key removed', 'success');
-    },
-    onError: (err: Error) => {
-      showToast(err.message || 'Failed to remove Stripe key', 'error');
-    },
-  });
-
-  return {
-    keys: data?.keys ?? [],
-    isLoading,
-    error,
-    saveKey,
-    removeKey,
-  };
-}
