+++ packages/dashboard/src/features/payments/hooks/useRazorpayWebhook.ts
@@ -4,22 +4,15 @@ import {
   razorpayService,
   type GetRazorpayStatusResponse,
 } from '#features/payments/services/razorpay.service';
+import { razorpayQueryKeys } from '#features/payments/queryKeys';
 import { useToast } from '#lib/hooks/useToast';
-
-const RAZORPAY_STATUS_QUERY_KEY = ['payments', 'razorpay', 'status'];
-const getRazorpayWebhookSetupQueryKey = (environment: RazorpayEnvironment) => [
-  'payments',
-  'razorpay',
-  environment,
-  'webhook-setup',
-];
 
 export function useRazorpayWebhook() {
   const queryClient = useQueryClient();
   const { showToast } = useToast();
 
   const { data, isLoading, error } = useQuery<GetRazorpayStatusResponse>({
-    queryKey: RAZORPAY_STATUS_QUERY_KEY,
+    queryKey: razorpayQueryKeys.status,
     queryFn: () => razorpayService.getStatus(),
     staleTime: 30 * 1000,
   });
@@ -28,8 +21,8 @@ export function useRazorpayWebhook() {
     mutationFn: (environment: RazorpayEnvironment) =>
       razorpayService.rotateWebhookSecret(environment),
     onSuccess: async () => {
-      await queryClient.invalidateQueries({ queryKey: RAZORPAY_STATUS_QUERY_KEY });
-      await queryClient.invalidateQueries({ queryKey: ['payments', 'razorpay'] });
+      await queryClient.invalidateQueries({ queryKey: razorpayQueryKeys.status });
+      await queryClient.invalidateQueries({ queryKey: razorpayQueryKeys.all });
       showToast('Razorpay webhook secret rotated', 'success');
     },
     onError: (err: Error) => {
@@ -47,7 +40,7 @@ export function useRazorpayWebhook() {
 
 export function useRazorpayWebhookSetup(environment: RazorpayEnvironment, enabled: boolean) {
   return useQuery({
-    queryKey: getRazorpayWebhookSetupQueryKey(environment),
+    queryKey: razorpayQueryKeys.webhookSetup(environment),
     queryFn: () => razorpayService.getWebhookSetup(environment),
     enabled,
     staleTime: 30 * 1000,
